% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/story.R
\name{story}
\alias{story}
\title{Generates a story for the given expression}
\usage{
story(expr)
}
\arguments{
\item{expr}{(expression) The R expression to generate the graph for.}
}
\description{
Generates a story for the given expression
}
